third_party/blink/web_tests/platform/generic/fast/forms/shadow-tree-exposure-expected.txt

Test to make sure shadow nodes are not exposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getSelection().anchorNode is null
PASS getSelection().anchorOffset is 0
PASS getSelection().focusNode is null
PASS getSelection().focusOffset is 0
PASS getSelection().isCollapsed is true
PASS getSelection().rangeCount is 0
PASS getSelection().baseNode is null
PASS getSelection().baseOffset is 0
PASS getSelection().extentNode is null
PASS getSelection().extentOffset is 0
PASS getSelection().type is 'None'

Add an input element.

PASS getSelection().anchorNode is container
PASS getSelection().anchorOffset is 1
PASS getSelection().focusNode is container
PASS getSelection().focusOffset is 1
PASS getSelection().isCollapsed is true
PASS getSelection().rangeCount is 1
PASS getSelection().getRangeAt(0).startContainer is container
PASS getSelection().getRangeAt(0).startOffset is 1
PASS getSelection().getRangeAt(0).endContainer is container
PASS getSelection().getRangeAt(0).endOffset is 1
PASS getSelection().baseNode is container
PASS getSelection().baseOffset is 1
PASS getSelection().extentNode is container
PASS getSelection().extentOffset is 1
PASS getSelection().type is "Range"

Add a textarea element.

PASS getSelection().anchorNode is container
PASS getSelection().anchorOffset is 2
PASS getSelection().focusNode is container
PASS getSelection().focusOffset is 2
PASS getSelection().isCollapsed is true
PASS getSelection().rangeCount is 1
PASS getSelection().getRangeAt(0).startContainer is container
PASS getSelection().getRangeAt(0).startOffset is 2
PASS getSelection().getRangeAt(0).endContainer is container
PASS getSelection().getRangeAt(0).endOffset is 2
PASS getSelection().baseNode is container
PASS getSelection().baseOffset is 2
PASS getSelection().extentNode is container
PASS getSelection().extentOffset is 2
PASS getSelection().type is "Range"


PASS successfullyParsed is true

TEST COMPLETE
